tools/esc-validator/requirements.txt
# ESC Validator - Phase 1 Requirements
# Core PDF and image processing
pdfplumber>=0.10.0
pypdf>=3.17.0
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0

# OpenCV for image processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Text processing and matching
python-Levenshtein>=0.21.0

# Visualization and reporting
matplotlib>=3.7.0
